Add files via upload
 A few aesthetic and functional changes. Made the previous month days and next month days stan
d on their own with a light gray color. Now the next task is making the selection of days not be 
on two items at once. its a weird bug that came about because I was using selecting by grabbing t
he text from an item. 2 would pick day two but seeing as there are 2 2's(the other one being from
 the next month) the selected orangered box will show up twice.